requirements.txt
certifi==2020.6.20
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.32.0
idna==3.3
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mysql-connector-python==8.0.28
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.1.0
protobuf==3.20.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
torch==1.11.0
torch-cluster==1.6.0
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.13
torch-spline-conv==1.2.1
tqdm==4.63.1
typing_extensions==4.1.1
urllib3==1.26.9
unidecode==1.3.6
